Group POS fetch analytics tracking in a single file
