+++ Yosemite/Yosemite/PointOfSale/POSItemFetchAnalyticsTracking.swift
@@ -6,4 +6,10 @@ public protocol POSItemFetchAnalyticsTracking {
     /// - Parameters:
     ///   - totalItems: The total number of items in the store
     func trackItemsFetchComplete(totalItems: Int)
+
+    /// Tracks when a remote search results fetch completes
+    /// - Parameters:
+    ///   - millisecondsSinceRequestSent: The time taken to fetch results in milliseconds
+    ///   - totalItems: The total number of items found in the search
+    func trackSearchRemoteResultsFetchComplete(millisecondsSinceRequestSent: Int, totalItems: Int)
 }
